requirements.txt

Flask==2.3.3
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.9.0
MarkupSafe==3.0.2


Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.22
greenlet==3.0.1
psycopg2-binary==2.9.9


Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0.post1
dnspython==2.8.0
idna==3.10


Flask-Login==0.6.3


python-dotenv==1.0.1
vercel-blob==0.4.2